Tweak Authomatic download to a specific commit
This is immediately prior for a change in Authomatic to use
python3-style `super()` calls with no arguments. This is a convenience
feature that unfortunately isn't working for us -- apparently because
the authomatic module is being reloaded for some reason. (This causes
class IDs to change, so super() fails with a class mismatch.)

Here's the problematic commit:
https://github.com/authomatic/authomatic/commit/3e52a9a3fe327490029e76979bb6aed7b80387e3